+++ libs/angular/config/src/lib/helpers/bootstrap-app-with-config.helper.ts
@@ -42,12 +42,15 @@ export async function bootstrapAppWithConfig<T>(
     }
   }
 
-  config = options?.getConfig
-    ? options?.getConfig(config as AngularConfig)
-    : config;
+  config = (
+    options?.getConfig ? options?.getConfig(config as AngularConfig) : config
+  ) as T;
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   (window as any)[ANGULAR_CONFIGURATION_SESSION_STORAGE] = config;
 
-  return platformBrowserDynamic().bootstrapModule(AppModule);
+  return {
+    module: platformBrowserDynamic().bootstrapModule(AppModule),
+    config,
+  };
 }