+++ libs/common/src/lib/dtos/index.ts
@@ -1,5 +1,0 @@
-export * from './bool-filter.dto';
-export * from './date-time-filter.dto';
-export * from './int-filter.dto';
-export * from './json-filter.dto';
-export * from './string-filter.dto';