+++ libs/common/src/lib/validator-transformer/transformers/transform-string-to-number.ts
@@ -1,0 +1,11 @@
+import { Transform, TransformFnParams } from 'class-transformer';
+
+export function TransformStringToNumber() {
+  return Transform(({ value }: TransformFnParams) => {
+    if (typeof value !== 'string') return value;
+
+    const number = Number(value);
+    if (Number.isNaN(number)) return value;
+    return number;
+  });
+}