+++ libs/common/src/lib/validator-transformer/validators/xor.ts
@@ -1,7 +1,0 @@
-import { Validate } from 'class-validator';
-
-import { XorConstraint } from '../constraints';
-
-export function XorValidate(constraints: string[]) {
-  return Validate(XorConstraint, constraints);
-}