+++ libs/nestjs/authentication/src/index.ts
@@ -1,7 +1,6 @@
 export * from './decorators';
 export * from './dtos';
 export * from './constants';
-export * from './enums';
 export * from './errors';
 export * from './guards';
 export * from './interfaces';