+++ libs/nestjs/authentication/src/interfaces/password-service.interface.ts
@@ -1,6 +1,0 @@
-export type RequestResetOptions = {
-  template?: number;
-  link?: string;
-  subject?: string;
-  variables?: Record<string, unknown>;
-};