+++ libs/nestjs/password/src/controllers/password.controller.ts
@@ -1,0 +1,24 @@
+import { Body, Controller, Patch } from '@nestjs/common';
+import { User } from '@prisma/client';
+
+import { PasswordUpdateDto } from '../dtos';
+import { PasswordService } from '../services';
+
+import { CurrentUser } from '@tractr/nestjs-core';
+
+@Controller('password')
+export class PasswordController {
+  constructor(private readonly passwordService: PasswordService) {}
+
+  @Patch('update')
+  async updatePassword(
+    @Body() { oldPassword, newPassword }: PasswordUpdateDto,
+    @CurrentUser() user: User,
+  ): Promise<void> {
+    await this.passwordService.updatePassword(
+      user.id,
+      oldPassword,
+      newPassword,
+    );
+  }
+}