+++ libs/nestjs/password/src/errors/index.ts
@@ -1,0 +1,2 @@
+export * from './bad-password.error';
+export * from './bad-reset-code.error';