+++ libs/terraform/service/admin/src/lib/admin.container.ts
@@ -1,6 +1,7 @@
 import { AdminContainerConfig } from './interfaces';
 
 import {
+  COMMON_SENSITIVE_ENVIRONMENT_VARIABLES,
   EnvironmentDefinition,
   HttpContainer,
 } from '@tractr/terraform-service-ecs';
@@ -25,6 +26,7 @@ export class AdminContainer extends HttpContainer<AdminContainerConfig> {
       name: 'ENVS_NAMES',
       value: this.getEnvNames()
         .filter((value) => value !== 'HTML_INDEX_PATH')
+        .filter((v) => !COMMON_SENSITIVE_ENVIRONMENT_VARIABLES.includes(v))
         .join(','),
     });