+++ libs/common/CHANGELOG.md
@@ -2,6 +2,151 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.62.0](https://github.com/tractr/stack/compare/v1.61.3...v1.62.0) (2022-10-12)
+
+
+
+## [1.61.3](https://github.com/tractr/stack/compare/v1.61.2...v1.61.3) (2022-10-12)
+
+
+
+## [1.61.2](https://github.com/tractr/stack/compare/v1.61.1...v1.61.2) (2022-10-06)
+
+
+
+## [1.61.1](https://github.com/tractr/stack/compare/v1.61.0...v1.61.1) (2022-10-05)
+
+
+
+# [1.61.0](https://github.com/tractr/stack/compare/v1.60.2...v1.61.0) (2022-09-29)
+
+
+### Features
+
+* **common:** add waitFor operator ([#506](https://github.com/tractr/stack/issues/506)) ([3af71b1](https://github.com/tractr/stack/commit/3af71b1c8540230e8c9d5639a30972236a5cb6b0))
+
+
+
+## [1.60.2](https://github.com/tractr/stack/compare/v1.60.1...v1.60.2) (2022-09-22)
+
+
+
+## [1.60.1](https://github.com/tractr/stack/compare/v1.60.0...v1.60.1) (2022-09-21)
+
+
+### Bug Fixes
+
+* update getConcatValueByPath for prisma query ([#510](https://github.com/tractr/stack/issues/510)) ([c9be995](https://github.com/tractr/stack/commit/c9be995fab130aaaf8b1e530134f4f7dcfb78c7a))
+
+
+
+# [1.60.0](https://github.com/tractr/stack/compare/v1.59.0...v1.60.0) (2022-09-19)
+
+
+
+# [1.59.0](https://github.com/tractr/stack/compare/v1.58.7...v1.59.0) (2022-09-19)
+
+
+
+## [1.58.7](https://github.com/tractr/stack/compare/v1.58.6...v1.58.7) (2022-09-19)
+
+
+
+## [1.58.6](https://github.com/tractr/stack/compare/v1.58.5...v1.58.6) (2022-09-13)
+
+
+
+## [1.58.5](https://github.com/tractr/stack/compare/v1.58.4...v1.58.5) (2022-09-13)
+
+
+
+## [1.58.4](https://github.com/tractr/stack/compare/v1.58.3...v1.58.4) (2022-09-08)
+
+
+
+## [1.58.3](https://github.com/tractr/stack/compare/v1.58.2...v1.58.3) (2022-09-08)
+
+
+
+## [1.58.2](https://github.com/tractr/stack/compare/v1.58.1...v1.58.2) (2022-09-07)
+
+
+
+## [1.58.1](https://github.com/tractr/stack/compare/v1.58.0...v1.58.1) (2022-09-06)
+
+
+
+# [1.58.0](https://github.com/tractr/stack/compare/v1.57.2...v1.58.0) (2022-09-02)
+
+
+
+## [1.57.2](https://github.com/tractr/stack/compare/v1.57.1...v1.57.2) (2022-09-01)
+
+
+
+## [1.57.1](https://github.com/tractr/stack/compare/v1.57.0...v1.57.1) (2022-08-31)
+
+
+
+# [1.57.0](https://github.com/tractr/stack/compare/v1.56.1...v1.57.0) (2022-08-31)
+
+
+### Features
+
+* split nestjs password and authentication management ([#497](https://github.com/tractr/stack/issues/497)) ([fe2ed5e](https://github.com/tractr/stack/commit/fe2ed5e0b71c115153d3bdcadd652c196dd9308f))
+
+
+
+## [1.56.1](https://github.com/tractr/stack/compare/v1.56.0...v1.56.1) (2022-08-26)
+
+
+
+# [1.56.0](https://github.com/tractr/stack/compare/v1.55.0...v1.56.0) (2022-08-19)
+
+
+
+# [1.55.0](https://github.com/tractr/stack/compare/v1.54.5...v1.55.0) (2022-08-18)
+
+
+
+## [1.54.5](https://github.com/tractr/stack/compare/v1.54.4...v1.54.5) (2022-08-18)
+
+
+
+## [1.54.4](https://github.com/tractr/stack/compare/v1.54.3...v1.54.4) (2022-08-17)
+
+
+
+## [1.54.3](https://github.com/tractr/stack/compare/v1.54.2...v1.54.3) (2022-08-17)
+
+
+
+## [1.54.2](https://github.com/tractr/stack/compare/v1.54.1...v1.54.2) (2022-08-17)
+
+
+
+## [1.54.1](https://github.com/tractr/stack/compare/v1.54.0...v1.54.1) (2022-08-17)
+
+
+
+# [1.54.0](https://github.com/tractr/stack/compare/v1.53.1...v1.54.0) (2022-08-17)
+
+
+### Features
+
+* update angular, nestjs, prisma and react-admin ([#468](https://github.com/tractr/stack/issues/468)) ([8122574](https://github.com/tractr/stack/commit/8122574544697a29a2acf1aef799fc568fb50be0))
+
+
+
+## [1.53.1](https://github.com/tractr/stack/compare/v1.53.0...v1.53.1) (2022-08-16)
+
+
+### Bug Fixes
+
+* add test coverage to 80% ([#446](https://github.com/tractr/stack/issues/446)) ([04f1857](https://github.com/tractr/stack/commit/04f18577cf177db6024bd3536aa8e7f6fb6199d8))
+
+
+
 # [1.53.0](https://github.com/tractr/stack/compare/v1.52.2...v1.53.0) (2022-08-15)